% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ris.R
\name{convert_ris}
\alias{convert_ris}
\title{Convert an RIS file to CSV format}
\usage{
convert_ris(in_file, out_file)
}
\arguments{
\item{in_file}{RIS file to process}

\item{out_file}{name of CSV file to write. This is passed to
\code{\link[base]{system2}} as the \code{stdout} parameter, so set
\code{out_file=T} to get the output back as the function return value.}
}
\value{
return value from \code{\link[base]{system2}}.
}
\description{
Deprecated. Use \code{\link{read_ris}} unless the speed gain is important.
Convert RIS to CSV using a Python script; the output can be loaded with
\code{\link{read_ris_csv}}.  Note that this generates a \emph{wide}-format
rather than long-format data frame.
}

