% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjects.R
\name{is_author}
\alias{is_author}
\title{Detect author names among subject headings}
\usage{
is_author(x)
}
\arguments{
\item{x}{character vector of subject terms}
}
\value{
logical vector indicating whether the corresponding element of
  \code{x} is an author term.
}
\description{
Given a vector of subject terms, determine which are (likely) author terms.
}
\details{
The heuristic is simply to look for a (possibly open-ended) birth-death date
range in the term. Note that this is highly imperfect, since books and
periodicals often also have date ranges. I have not found an alternative to
removing these by hand.
}
\examples{
is_author("Joyce, James (1882-1941)")

is_author("South Asian literature")

# but N.B.
is_author("A la recherche du temps perdu (1913-1929)")

}
\seealso{
\code{\link{subjects_frame}}, \code{\link{subject_author}}
}
