% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlaib-package.R
\docType{package}
\name{mlaib}
\alias{mlaib}
\alias{mlaib-package}
\title{Explore MLA bibliography data}
\description{
This package gives some rudimentary utilities for loading and preprocessing
MLA International Bibliography data or other RIS-formatted bibliographic
data.
}

