% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlaib-package.R
\docType{package}
\name{mlaibr}
\alias{mlaibr}
\alias{mlaibr-package}
\title{Explore MLA bibliography data}
\description{
This package gives some rudimentary utilities for loading and preprocessing
MLA International Bibliography data or other RIS-formatted bibliographic
data.
}
