% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-parsing.R
\name{N1_field}
\alias{N1_field}
\title{Extract finer-grained metadata from the note field}
\usage{
N1_field(x, field)
}
\arguments{
\item{x}{character vector}

\item{field}{name of field to extract (without colon). Case sensitive.}
}
\value{
character vector of extracted values. Validate carefully. Missing
  values are represented as \code{NA}.
}
\description{
In the MLA bibliography, the N1 (note) field contains further metadata which
is general consistently identified and can be programmatically extracted.
Examples include the \code{Accession Number}, \code{Publication Type},
\code{DOI}, and \code{Peer Reviewed} fields.
}
\details{
Fields are reliably delimited only by a terminating period, and there is
sometimes a missing period at the end of a field, e.g. in MLAIB acc nos.
1998070296, 1979107484, 1979202534. If the field data contains a period this
is a problem, but the only case in which I know this to happen is in
\code{DOI} fields. This is treated as a special case.
}
\examples{
\dontrun{
b <- read_ris("bib.ris") |>
    spread_ris()

# Extract peer reviewing flag
# N.B. This information is very incomplete in the MLAIB
b |>
    mutate(peer=N1_field(N1, "Peer Reviewed") == "Yes")

# Extract publication type
# This is not just a recoding of the RIS field TY. Cross-tabulate.
b |>
    mutate(pubtype=N1_field(N1, "Publication Type")) |>
    count(TY, pubtype)
}

}
