% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ris.R
\name{spread_ris}
\alias{spread_ris}
\title{Convert bibliographic records from long to wide format}
\usage{
spread_ris(x, multi_sep = ";;")
}
\arguments{
\item{x}{data frame with \code{id,field,value} columns}

\item{multi_sep}{separator string for repeated fields. Should not occur
within field values.}
}
\value{
data frame with one row for each bibliographic item record and one
  column for each field. Missing values are represented with \code{NA}.
}
\description{
The data frame from \code{\link{read_ris}} has many rows for each
bibliographic record. Many questions relate to bibliographic items as a unit,
in which case it is convenient to have a data frame in the corresponding wide
form. However, note that this is a somewhat non-normalized format since many
RIS fields can repeat within a record. Repeat fields are joined into a single
string value.
}
\details{
It would be a little more formal to produce list columns from repeating
fields, but that's just a drag. It may make more sense to normalize by hand
by making separate tables of the repeating columns and keying to ID (or
accession number), then spreading out only the single columns of interest
here.
}
\seealso{
\code{\link{read_ris}}
}
