% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subjects.R
\name{strip_subject_relation}
\alias{strip_subject_relation}
\title{Remove relational phrases from subject headings}
\usage{
strip_subject_relation(x, rels = getOption("mlaib.relations"))
}
\arguments{
\item{x}{character vector of headings}

\item{rels}{character vector of regular expressions matching relation terms.
  The anchor \code{^} is added at the start and then the vector is collapsed
  using the alternator \code{|} as a separator. A default list is set as
  package option \code{mlaib.relations}, but this was found by trial and
  error and may need modification.}
}
\value{
the headings with relation terms removed
}
\description{
Many subject headings specify a relation to a term: not just "Joyce, James"
but "compared to Joyce, James." For aggregating purposes, these relations can
often be ignored. This function attempts to remove these relational phrases
and leave only the main terms. Subheadings (delimited by a colon) are also
removed.
}
\details{
Since the relation terms can be composed ("compared to treatment of"), the
function repeatedly strips the terms, up to a maximum number of iterations
given by the package option \code{mlaib.rel_strip_iterations}. It issues a
warning if relation terms are still present at the end.
}

