% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjects.R
\name{subject_author}
\alias{subject_author}
\title{Convert subject terms for authors to names alone}
\usage{
subject_author(x)
}
\arguments{
\item{x}{character vector of subject terms}
}
\value{
character vector of author terms
}
\description{
Given subject terms containing author names, strip away date ranges to yield
\code{"Last, First Middle Middle"} names. No further de-duplication is
performed. Subheadings and relation terms are assumed already stripped (as by
\code{\link{strip_subject_relation}}).
}
\details{
Note that removing date ranges potentially also removes a disambiguation
among authors with the same name. On your head be it.
}
\examples{
subject_author("Joyce, James (1882-1941)")

}
\seealso{
\code{\link{subjects_frame}}, \code{\link{is_author}}
}

