% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subjects.R
\name{subject_authors_frame}
\alias{subject_authors_frame}
\title{Filter a subjects frame down to author-subjects}
\usage{
subject_authors_frame(x, non_authors = NULL)
}
\arguments{
\item{x}{data frame with author terms in a \code{value} column}

\item{non_authors}{(optional) character vector of names to reject that
  otherwise pass the filter.}
}
\value{
data frame like \code{x} but only rows where \code{value} is an
  author name retained. You might wish to use \code{\link[dplyr]{distinct}}
  to deduplicate this.
}
\description{
Given the result form \code{\link{subjects_frame}}, remove all but the names
of authors. Any birth-death date ranges after names are removed. No further
de-duplication is performed.
}
\details{
Note that removing date ranges potentially also removes a disambiguation
among authors with the same name. On your head be it.
}
\seealso{
\code{\link{subjects_frame}}
}

