% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subjects.R
\name{subject_authors_last}
\alias{subject_authors_last}
\title{Author term to lowercase last name only}
\usage{
subject_authors_last(values)
}
\arguments{
\item{values}{character vector of \code{"Last, First ..."} terms}
}
\value{
character vector of \code{"last"} names
}
\description{
A convenience function for stripping down to an author keyword.
}

