% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subjects.R
\name{subjects_frame}
\alias{subjects_frame}
\title{Generate a data frame of the id-subject-heading table}
\usage{
subjects_frame(bib, rels = getOption("mlaib.relations"))
}
\arguments{
\item{bib}{long-format table of RIS information}

\item{rels}{character vector of relation patterns to pass to
  \code{\link{strip_subject_relation}}}
}
\value{
data frame with \code{id,value} columns
}
\description{
It's often convenient to have a data frame with just subject terms and id's
that can be filtered and then joined back onto a table of bibliographic
items. This function is just a wrapper for filtering only \code{KW} fields
and then applying \code{\link{strip_subject_relation}}.
}
\seealso{
\code{\link{subject_authors_frame}}
}

