% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-parsing.R
\name{Y1_year}
\alias{Y1_year}
\title{Naive parsing of dates into years}
\usage{
Y1_year(x)
}
\arguments{
\item{x}{character vector (Y1 field)}
}
\value{
vector of \code{Date}s corresponding to years
}
\description{
The \code{Y1} field holds the publication date information. This normally
contains a year, and sometimes also contains month, day, or season
information. This function extracts just the year and recodes it as a
\code{Date} on the first day of the year.
}
\examples{
\dontrun{
# To get at the messy finer-grained information, try:
str_split_fixed(x, coll("/"), 4)
}

}
