% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ris.R
\name{read_ris}
\alias{read_ris}
\title{Read RIS files into a long-format data frame}
\usage{
read_ris(filenames, fields = getOption("mlaib.ris_keep"), src_labels = NULL)
}
\arguments{
\item{filenames}{vector of names of RIS files to read. Can also be a list of
connections, or a single connection. As a convenience, if a name ends in
"\code{.zip}," the file will be assumed to be a zip archive containing a
single RIS file (which is the format supplied in EBSCOhost exports).}

\item{fields}{which RIS fields to keep in the result. A default list is set
by the package option \code{mlaib.ris_keep}. To keep all fields, set
\code{fields=NULL}.}
}
\value{
a data frame with columns \code{id,field,value}
}
\description{
This function loads one or more RIS files into a long data frame. It assigns
unique record ID numbers and optionally filters out unwanted fields. The
result has one line for each record-field combination, including possible
repeats when a record has multiple instances of a field. The result is not
quite "tidy," because no special parsing is done to extract information
within the catch-all \code{N1} field; use \code{\link{N1_field}} for that.
Note also that all fields remain in string format; for example, dates are not
parsed (see \code{\link{Y1_year}}). To project to a data frame with one row
for each bibliographic record, use \code{\link{spread_ris}}.
}
\seealso{
\code{\link{spread_ris}}
}

